requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-dotenv>=0.19.0
openai>=1.0.0
google-auth>=2.3.0
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
python-telegram-bot>=13.7
httpx>=0.23.0
python-multipart>=0.0.5
requests>=2.28.0
